feat #84 商家账户注册（1）

+++ muyuan-product/src/main/java/com/muyuan/product/domains/service/impl/GoodsCategoryDomainServiceImpl.java
@@ -1,10 +1,18 @@
 package com.muyuan.product.domains.service.impl;
 
 import com.muyuan.common.core.constant.GlobalConst;
+import com.muyuan.common.core.constant.RedisConst;
+import com.muyuan.common.core.enums.ResponseCode;
+import com.muyuan.common.core.exception.MuyuanException;
 import com.muyuan.common.redis.manage.RedisCacheService;
+import com.muyuan.product.domains.dto.CategoryAttributeDTO;
 import com.muyuan.product.domains.dto.GoodsCategoryDTO;
+import com.muyuan.product.domains.model.BrandCategory;
+import com.muyuan.product.domains.model.CategoryAttribute;
 import com.muyuan.product.domains.model.GoodsCategory;
+import com.muyuan.product.domains.repo.CategoryAttributeRepo;
 import com.muyuan.product.domains.repo.GoodsCategoryRepo;
+import com.muyuan.product.domains.service.CategoryAttributeDomainService;
 import com.muyuan.product.domains.service.GoodsCategoryDomainService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,6 +39,8 @@ import java.util.stream.Collectors;
 public class GoodsCategoryDomainServiceImpl implements GoodsCategoryDomainService {
 
     private GoodsCategoryRepo goodsCategoryRepo;
+
+    private CategoryAttributeRepo categoryAttributeRepo;
 
     private RedisCacheService redisCacheService;
 
@@ -62,7 +72,7 @@ public class GoodsCategoryDomainServiceImpl implements GoodsCategoryDomainServic
         goodsCategory.initRoot(rootCount);
         goodsCategory.save(goodsCategoryRepo);
         goodsCategory.setAncestors(String.valueOf(goodsCategory.getId()));
-        goodsCategory.update(goodsCategoryRepo,"ancestors");
+        goodsCategory.update(goodsCategoryRepo, "ancestors");
     }
 
     private void addSubNodeCategory(GoodsCategory goodsCategory, GoodsCategory parent) {
@@ -82,6 +92,7 @@ public class GoodsCategoryDomainServiceImpl implements GoodsCategoryDomainServic
     public void update(GoodsCategoryDTO goodsCategoryDTO) {
         GoodsCategory goodsCategory = goodsCategoryDTO.convert();
         goodsCategory.save(goodsCategoryRepo);
+        deleteCategoryCache(goodsCategory);
     }
 
     @Override
@@ -105,7 +116,23 @@ public class GoodsCategoryDomainServiceImpl implements GoodsCategoryDomainServic
     @Override
     public Optional<GoodsCategory> detail(GoodsCategory goodsCategory) {
 
-        GoodsCategory category = goodsCategoryRepo.selectDetail(goodsCategory);
+        GoodsCategory category = redisCacheService.getAndUpdate(CATEGORY_KEY_PREFIX + goodsCategory.getCode(), () ->
+                goodsCategoryRepo.selectOne(GoodsCategory.builder()
+                        .code(goodsCategory.getCode())
+                        .build())
+        ,GoodsCategory.class);
+
+        if (ObjectUtils.isEmpty(category)) {
+            return Optional.empty();
+        }
+
+        List<CategoryAttribute> attributes = redisCacheService.getAndUpdateList(CategoryAttributeDomainService.CATEGORY_ATTRIBUTE_KEY_PREFIX + goodsCategory.getCode(), () ->
+                        categoryAttributeRepo.select(CategoryAttributeDTO.builder()
+                                .categoryCode(goodsCategory.getCode())
+                                .build())
+                , CategoryAttribute.class);
+
+        category.setAttributes(attributes);
 
         return Optional.ofNullable(category);
     }
@@ -119,16 +146,29 @@ public class GoodsCategoryDomainServiceImpl implements GoodsCategoryDomainServic
 
         Long[] parentIds = ids;
         List<GoodsCategory> sub;
+
+        List<GoodsCategory> list = goodsCategoryRepo.list(GoodsCategoryDTO.builder().ids(ids).build());
+        List<Long> codes = list.stream().map(GoodsCategory::getCode).collect(Collectors.toList());
         do {
-            /**
-             * TODO:判断是否有商品关联分类
-             */
             sub = goodsCategoryRepo.list(GoodsCategoryDTO.builder().parentIds(parentIds).build());
             if (!sub.isEmpty()) {
                 toDelete.addAll(sub.stream().map(GoodsCategory::getId).collect(Collectors.toList()));
+                codes.addAll(sub.stream().map(GoodsCategory::getCode).collect(Collectors.toList()));
             }
         } while (!sub.isEmpty());
 
+        List<BrandCategory> brandCategories = goodsCategoryRepo.selectBrand(codes.toArray(new Long[0]));
+        if (ObjectUtils.isNotEmpty(brandCategories)) {
+            throw new MuyuanException(ResponseCode.FAIL.getCode(), "有品牌关联当前分类");
+        }
+
         goodsCategoryRepo.delete(toDelete.toArray(new Long[0]));
+
+        redisCacheService.del(CATEGORY_KEY_PREFIX + RedisConst.ALL_PLACE_HOLDER);
+    }
+
+
+    private void deleteCategoryCache(GoodsCategory goodsCategory) {
+        redisCacheService.del(CATEGORY_KEY_PREFIX + goodsCategory);
     }
 }
